Introduce the lambda and p refinement
The parameters lambda and the power p are refined at every iterations
This improves greatly the convergence speed and the robustess
Make the pyramid size reproductible
Use only the PSF estimation zone to compute the covariance : more efficient and reliable
Add extra security padding on images to prevent boundaries problems
Compact the code for the API
Rewrite some options to make them more user-friendy
Change the mask input : now, input the center and the size of the mask is computed accordingly to
the blur size.
Various minor corrections
More comments.